third_party/blink/web_tests/virtual/third-party-storage-partitioning/fast/filesystem/simple-persistent-expected.txt

requestFileSystem PERSISTENT test.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Successfully obtained PERSISTENT FileSystem:file__0:Persistent
PASS fileSystem.name.length > 0 is true
PASS fileSystem.root.fullPath is "/"
PASS successfullyParsed is true

TEST COMPLETE
